+++ src/Context.php
@@ -9,6 +9,8 @@
 
 namespace PhpMyAdmin\SqlParser;
 
+use PhpMyAdmin\SqlParser\Exceptions\LoaderException;
+
 /**
  * Holds the configuration of the context that is currently used.
  *
@@ -443,7 +445,7 @@ abstract class Context
      * @param string $context name of the context or full class name that
      *                        defines the context
      *
-     * @throws \Exception if the specified context doesn't exist
+     * @throws LoaderException if the specified context doesn't exist
      */
     public static function load($context = '')
     {
@@ -455,8 +457,9 @@ abstract class Context
             $context = self::$contextPrefix . $context;
         }
         if (!class_exists($context)) {
-            throw new \Exception(
-                'Specified context ("' . $context . '") does not exist.'
+            throw new LoaderException(
+                'Specified context ("' . $context . '") does not exist.',
+                $context
             );
         }
         self::$loadedContext = $context;
@@ -491,7 +494,7 @@ abstract class Context
             try {
                 // Trying to load the new context.
                 static::load($context);
-            } catch (\Exception $e) {
+            } catch (LoaderException $e) {
                 // If it didn't work, we are looking for a new one and skipping
                 // over to the next generation that will try the new context.
                 $context = preg_replace(