+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Use custom LoaderException for context loading errors.
+
 ## [4.1.9] - 2017-07-12
 
 * Various code cleanups.